refactor: 서비스 코드 수정1
