+++ src/main/java/com/example/oatnote/memoTag/service/relation/MemoTagRelationRepository.java
@@ -1,10 +1,10 @@
-package com.example.oatnote.memoTag.service.memoTagRelation;
+package com.example.oatnote.memoTag.service.relation;
 
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.example.oatnote.memoTag.service.memoTagRelation.models.MemoTagRelation;
+import com.example.oatnote.memoTag.service.relation.models.MemoTagRelation;
 
 public interface MemoTagRelationRepository extends MongoRepository<MemoTagRelation, String> {
